state: move trie walking methods to PortalTrie class
